Add all models - see detail
All models are coded, but only 2 non-dependant models are not commented
out.
Initial test data for the Repertoire model was deleted.
Values for the Section model were input manually.

This is an attempt to avoid ForeignKey errors as dependant data is added
to the database.

Next up: manually adding Repertoire data.
	modified:   assets/images/string-rota-erd-schema.png
	modified:   string_rota/admin.py
	new file:   string_rota/migrations/0006_auto_20221216_1740.py
	new file:   string_rota/migrations/0007_alter_section_section_name.py
	modified:   string_rota/models.py