Modify models - see details
The Rota model was deleted and the Djsngo Groups functiality was used
instead.
The Player model player_name field was changed to player_first_name,
and additional field of player_last_name was added, and a function to
create a full name (player_name) was added. This matches the structure
of the PLayer data to be imported.

	modified:   assets/images/string-rota-erd-schema.png
	modified:   string_rota/admin.py
	new file:   string_rota/migrations/0008_player.py
	modified:   string_rota/models.py
	modified:   string_rota/views.py