Change Player_Performance default - see detail
The performance_status default was changed to 'Not Available'
in Player_Project.
Add setup_player_project() to setup_tables.py.
This creates a Player_Project record for every player in every project.

As the rota-manager for seating plan allocates a player to a project,
this Player_Project record will also be updated.
	renamed:    load_data.py -> set_up.py
	new file:   string_rota/migrations/0038_auto_20221221_0957.py
	new file:   string_rota/migrations/0039_alter_player_project_performance_status.py
	modified:   string_rota/models.py
	renamed:    utilities/load_data_main.py -> utilities/load_rsno_data.py
	new file:   utilities/setup_tables.py