Updated settings and requirements
This project was transfered to the Gitpod
CI Student account after running out of Gitpod hours.
Django needed to be reinstalled.

Settings.py were updated so that:
- The local database was used if in DEVELPMENT. If not, then
the deployed Elephantsql database was used
- For Static and media files, the default django local settings
were used if in DEVELOPMENT. If deployed (not in DEVELOPMENT),
the cloudinary settings were used.
On prior project submission, the static files were not accesed correctly
on deployment. We shall see if they are correctly accesed now,