[#138] Add register user functionality
POST method of register class recieves:
- the selected player to register id,
- username
- password1
- password2
It checkes the passwords are the same, then sets valid to true
and creates a new user.
The new user is given the first and last name of the Player record.
and added to the group 'Player'
The Player.users_django field is linked to the user

Success and advisory messeges are passed to the redirect page.
	modified:   player_info/views.py

Add UserCreationForm to register.html
	modified:   player_info/templates/player_info/register.html